requirements.txt
psycopg2-binary
sqlalchemy        # optional but nice
python-dotenv     # optional for environment variables